move gyrofrequency tests to plasmapy/formulary/tests/test_frequencies.py
